feat: delete FootRight
